+++ doc/changelog.rst
@@ -12,6 +12,7 @@ Unreleased
 Core
 ~~~~
 
+- Fix bug: Prevent duplicate categories from all_visible-filter
 - Add support for using pricing templatetags for services
 - Make refund creation atomic
 - Allow refund only for non editable orders